pyproject.toml
[project]
name = "crypto-trading-bot"
version = "1.0.0"
description = "AI-powered crypto trading bot with Telegram notifications"
requires-python = ">=3.10"
dependencies = [
    "supabase>=2.27.0",
    "openai>=1.59.0",
    "ccxt>=4.4.0",
    "python-telegram-bot>=20.7",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W"]
